requirements.txt
# Core Dependencies
torch>=1.10.0
torch-geometric>=2.0.0
torch-scatter
torch-sparse
torch-cluster

# Scientific Computing
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
scipy>=1.7.0

# Chemistry
rdkit>=2022.03.1
openbabel-wheel  # For MOL2 file handling

# Configuration & Utilities
pyyaml>=5.4.0
easydict>=1.9
tqdm>=4.60.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Tensor board (for training)
tensorboard>=2.8.0

# Testing (required for running tests)
pytest>=6.2.0
